requirements.txt
Django==4.2.7
djangorestframework==3.14.0
mysqlclient==2.2.0
requests==2.31.0
Pillow==10.0.1
python-dotenv==1.0.0
django-filter==23.3

